refactor: remove unneeded activate route from device pages
