+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3'
-
 services:
   openamtui:
     image: openamtui
