+++ src/app/app-routing.module.ts
@@ -18,14 +18,12 @@ import { WirelessDetailComponent } from './wireless/wireless-detail/wireless-det
 import { WirelessComponent } from './wireless/wireless.component'
 import { IEEE8021xComponent } from './ieee8021x/ieee8021x.component'
 import { IEEE8021xDetailComponent } from './ieee8021x/ieee8021x-detail/ieee8021x-detail.component'
-import { ExplorerComponent } from './explorer/explorer.component'
 import { MonacoEditorModule } from 'ngx-monaco-editor-v2'
 
 const routes: Routes = [
   { path: 'devices', component: DevicesComponent },
   { path: 'devices/:id', component: DeviceDetailComponent },
   { path: 'devices/:id/:component', component: DeviceDetailComponent },
-  { path: 'devices/:id/explorer', component: ExplorerComponent },
   { path: 'profiles', component: ProfilesComponent },
   {
     path: 'profiles/new',
