+++ src/app/devices/device-detail/device-detail.component.ts
@@ -27,7 +27,7 @@ import { MatTabGroup, MatTab } from '@angular/material/tabs'
 import { MatSidenavContainer, MatSidenav, MatSidenavContent } from '@angular/material/sidenav'
 import { DeviceToolbarComponent } from '../device-toolbar/device-toolbar.component'
 import { ActivatedRoute, Router, RouterLink, RouterLinkActive } from '@angular/router'
-import { ExplorerComponent } from 'src/app/explorer/explorer.component'
+import { ExplorerComponent } from '../explorer/explorer.component'
 import { AlarmsComponent } from '../alarms/alarms.component'
 import { CertificatesComponent } from '../certificates/certificates.component'
 import { EventLogComponent } from '../event-log/event-log.component'
