+++ src/app/devices/event-log/event-log.component.spec.ts
@@ -138,6 +138,7 @@ describe('EventLogComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(EventLogComponent)
     component = fixture.componentInstance
+    component.deviceId = 'guid'
     fixture.detectChanges()
   })
