+++ src/app/devices/explorer/explorer.component.html
